+++ src/payload/index.ts
@@ -4,3 +4,8 @@ export {
 } from './createPayloadConfig'
 export { createGetPayload } from './createGetPayload'
 export { resolveD1Binding } from './resolveD1Binding'
+export {
+  createMockPayload,
+  type MockFixtures,
+  type MockPayload,
+} from './mockPayload'